Fixed job i/o.

+++ internal/web/web.go
@@ -87,7 +87,7 @@ func deleteJobHandler(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	} else {
-		_, err := db.DeleteJob(id)
+		_, err := db.DeleteJob(int64(id))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return
@@ -116,7 +116,7 @@ func jobHandler(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-	job, err := db.GetJob(id)
+	job, err := db.GetJob(int64(id))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
